Moved EF Entities into "Schema" namespaces
