+++ src/AdminAssistant.DAL/EntityFramework/ApplicationDbContext .cs
@@ -3,6 +3,11 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AdminAssistant.DAL.EntityFramework.Model;
+using AdminAssistant.DAL.EntityFramework.Model.Accounts;
+using AdminAssistant.DAL.EntityFramework.Model.AssetRegister;
+using AdminAssistant.DAL.EntityFramework.Model.Budget;
+using AdminAssistant.DAL.EntityFramework.Model.Contacts;
+using AdminAssistant.DAL.EntityFramework.Model.Documents;
 using AdminAssistant.DomainModel.Infrastructure;
 using AdminAssistant.DomainModel.Modules.Accounts;
 using Microsoft.EntityFrameworkCore;
