+++ src/AdminAssistant.DAL/EntityFramework/Model/Accounts/BankAccountEntity.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Accounts
 {
     public class BankAccountEntity
     {
