+++ src/AdminAssistant.DAL/EntityFramework/Model/Accounts/BankAccountTransactionEntity.cs
@@ -2,7 +2,7 @@ using System;
 using AdminAssistant.DomainModel.Modules.Accounts;
 using AdminAssistant.Framework.TypeMapping;
 
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Accounts
 {
     public class BankAccountTransactionEntity : IMapFrom<BankAccountTransaction>, IMapTo<BankAccountTransaction>
     {
