+++ src/AdminAssistant.DAL/EntityFramework/Model/Accounts/BankAccountTypeEntity.cs
@@ -1,7 +1,7 @@
 using AdminAssistant.DomainModel.Modules.Accounts;
 using AdminAssistant.Framework.TypeMapping;
 
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Accounts
 {
     public class BankAccountTypeEntity : IMapTo<BankAccountType>
     {
