+++ src/AdminAssistant.DAL/EntityFramework/Model/AuditEntity.cs
@@ -1,4 +1,7 @@
 using System;
+using AdminAssistant.DAL.EntityFramework.Model.Accounts;
+using AdminAssistant.DAL.EntityFramework.Model.AssetRegister;
+using AdminAssistant.DAL.EntityFramework.Model.Contacts;
 
 namespace AdminAssistant.DAL.EntityFramework.Model
 {
