+++ src/AdminAssistant.DAL/EntityFramework/Model/Documents/DocumentEntity.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
-
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Documents
 {
     public class DocumentEntity
     {
